First Draft of MakeCall Class Created Big Labels in NewCallList class and MakeCalls class that takes up the entire form and is brought to the front so user can't push buttons or edit data when saving or creating files or when Browser is starting or user is logging into SpoofCard
finalized the first draft of MakeCalls class.
MakeCalls class starts SpoofCard and gets the user to log into it. Then it iterates through every employee and displays their data. It also creates a new SpoofCard Pin for each employee. And it keeps track of the amount of engagements throughout the project and the number of today's engagements.